:sparkles: Update API and tests — [#437]
