+++ src/sdg/api/serializers/core.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from sdg.api.serializers.organisaties import OrganisatieBaseSerializer
+from sdg.api.serializers.organisaties import LokaleOverheidBaseSerializer
 from sdg.api.serializers.producten import ProductBaseSerializer
 from sdg.core.models import ProductenCatalogus
 
@@ -8,7 +8,7 @@ from sdg.core.models import ProductenCatalogus
 class ProductenCatalogusSerializer(serializers.HyperlinkedModelSerializer):
     """Serializer for ProductenCatalogus details, including organizations and products."""
 
-    organisatie = OrganisatieBaseSerializer(source="lokale_overheid")
+    organisatie = LokaleOverheidBaseSerializer(source="lokale_overheid")
     producten = ProductBaseSerializer(many=True)
 
     class Meta:
