+++ src/sdg/api/tests/test_locaties.py
@@ -1,0 +1,30 @@
+from rest_framework import status
+from rest_framework.reverse import reverse
+from rest_framework.test import APITestCase
+
+from sdg.organisaties.tests.factories.overheid import LocatieFactory
+
+
+class LocatiesTests(APITestCase):
+    def test_list_locations(self):
+        LocatieFactory.create_batch(2)
+        list_url = reverse("api:locatie-list")
+
+        response = self.client.get(list_url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.json()["results"]
+        self.assertEqual(2, len(data))
+
+    def test_retrieve_location_by_uuid(self):
+        locatie = LocatieFactory.create()
+        detail_url = reverse("api:locatie-detail", args=[locatie.uuid])
+
+        response = self.client.get(detail_url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.json()
+
+        self.assertEqual(str(locatie.uuid), data["uuid"])
