+++ src/sdg/api/tests/test_organisaties.py
@@ -1,0 +1,79 @@
+from rest_framework import status
+from rest_framework.reverse import reverse
+from rest_framework.test import APITestCase
+
+from sdg.organisaties.tests.factories.overheid import (
+    LocatieFactory,
+    LokaleOverheidFactory,
+)
+
+
+class OrganisatiesTests(APITestCase):
+    def test_list_organizations(self):
+        LokaleOverheidFactory.create_batch(2)
+        list_url = reverse("api:lokaleoverheid-list")
+
+        response = self.client.get(list_url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.json()["results"]
+        self.assertEqual(len(data), 2)
+
+    def test_retrieve_organization_by_uuid(self):
+        municipality = LokaleOverheidFactory.create()
+        detail_url = reverse("api:lokaleoverheid-detail", args=[municipality.uuid])
+
+        response = self.client.get(detail_url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.json()
+
+        self.assertEqual(
+            {
+                "url": f"http://testserver{reverse('api:lokaleoverheid-detail', args=[municipality.uuid])}",
+                "uuid": str(municipality.uuid),
+                "owmsIdentifier": municipality.organisatie.owms_identifier,
+                "owmsPrefLabel": municipality.organisatie.owms_pref_label,
+                "owmsEndDate": municipality.organisatie.owms_end_date.isoformat(),
+                "locaties": [],
+                "catalogi": [],
+                "contactNaam": municipality.contact_naam,
+                "contactWebsite": municipality.contact_website,
+                "contactEmailadres": municipality.contact_emailadres,
+                "contactTelefoonnummer": municipality.contact_telefoonnummer,
+                "bevoegdeOrganisaties": [],
+                "ondersteuningsOrganisatie": {
+                    "owmsIdentifier": municipality.ondersteunings_organisatie.owms_identifier,
+                    "owmsPrefLabel": municipality.ondersteunings_organisatie.owms_pref_label,
+                    "owmsEndDate": municipality.ondersteunings_organisatie.owms_end_date.isoformat(),
+                },
+            },
+            data,
+        )
+
+
+class LocatiesTests(APITestCase):
+    def test_list_locations(self):
+        LocatieFactory.create_batch(2)
+        list_url = reverse("api:locatie-list")
+
+        response = self.client.get(list_url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.json()["results"]
+        self.assertEqual(2, len(data))
+
+    def test_retrieve_location_by_uuid(self):
+        locatie = LocatieFactory.create()
+        detail_url = reverse("api:locatie-detail", args=[locatie.uuid])
+
+        response = self.client.get(detail_url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.json()
+
+        self.assertEqual(str(locatie.uuid), data["uuid"])
