+++ src/sdg/api/tests/test_producten.py
@@ -4,41 +4,13 @@ from rest_framework.reverse import reverse
 from rest_framework.test import APITestCase
 
 from sdg.core.tests.factories.catalogus import ProductenCatalogusFactory
-from sdg.organisaties.tests.factories.overheid import (
-    LocatieFactory,
-    LokaleOverheidFactory,
-)
+from sdg.organisaties.tests.factories.overheid import BevoegdeOrganisatieFactory
 from sdg.producten.tests.constants import NOW_DATE, PAST_DATE
 from sdg.producten.tests.factories.localized import LocalizedProductFactory
 from sdg.producten.tests.factories.product import (
     ReferentieProductFactory,
     ReferentieProductVersieFactory,
 )
-
-
-class CatalogiTests(APITestCase):
-    def test_list_catalogs(self):
-        ProductenCatalogusFactory.create_batch(2)
-        list_url = reverse("api:productencatalogus-list")
-
-        response = self.client.get(list_url)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        data = response.json()["results"]
-        self.assertEqual(2, len(data))
-
-    def test_retrieve_catalog_by_uuid(self):
-        catalog = ProductenCatalogusFactory.create()
-        detail_url = reverse("api:productencatalogus-detail", args=[catalog.uuid])
-
-        response = self.client.get(detail_url)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        data = response.json()
-
-        self.assertEqual(str(catalog.uuid), data["uuid"])
 
 
 class ProductenTests(APITestCase):
@@ -52,6 +24,28 @@ class ProductenTests(APITestCase):
 
         data = response.json()["results"]
         self.assertEqual(2, len(data))
+
+    def test_list_products_bevoegde_organisatie_missing_organisatie(self):
+        catalogus = ProductenCatalogusFactory.create()
+        bevoegde_organisatie = BevoegdeOrganisatieFactory.create(
+            lokale_overheid=catalogus.lokale_overheid,
+            organisatie=None,
+        )
+        ReferentieProductFactory.create(
+            catalogus=catalogus,
+            bevoegde_organisatie=bevoegde_organisatie,
+        )
+        list_url = reverse("api:product-list")
+
+        response = self.client.get(list_url)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.json()["results"][0]
+        self.assertEqual(
+            bevoegde_organisatie.naam, data["bevoegdeOrganisatie"]["owmsPrefLabel"]
+        )
+        self.assertIsNone(data["bevoegdeOrganisatie"]["owmsIdentifier"])
 
     @freeze_time(NOW_DATE)
     def test_retrieve_product_by_uuid(self):
@@ -73,6 +67,7 @@ class ProductenTests(APITestCase):
         self.assertEqual(f"{product.upn.upn_label}", data["upnLabel"])
         self.assertEqual(f"{product.upn.upn_uri}", data["upnUri"])
         self.assertEqual(product.product_valt_onder, data["productValtOnder"])
+        self.assertEqual(product.bevoegde_organisatie, data["bevoegdeOrganisatie"])
         self.assertEqual(1, data["versie"])
         self.assertEqual(
             f"http://testserver{reverse('api:productencatalogus-detail', args=[product.catalogus.uuid])}",
@@ -147,74 +142,3 @@ class ProductenTests(APITestCase):
         data = response.json()["results"]
         self.assertEqual(1, len(data))
         self.assertEqual(concept.versie, data[0]["versie"])
-
-
-class OrganisatiesTests(APITestCase):
-    def test_list_organizations(self):
-        LokaleOverheidFactory.create_batch(2)
-        list_url = reverse("api:lokaleoverheid-list")
-
-        response = self.client.get(list_url)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        data = response.json()["results"]
-        self.assertEqual(len(data), 2)
-
-    def test_retrieve_organization_by_uuid(self):
-        municipality = LokaleOverheidFactory.create()
-        detail_url = reverse("api:lokaleoverheid-detail", args=[municipality.uuid])
-
-        response = self.client.get(detail_url)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        data = response.json()
-
-        self.assertEqual(
-            {
-                "url": f"http://testserver{reverse('api:lokaleoverheid-detail', args=[municipality.uuid])}",
-                "uuid": str(municipality.uuid),
-                "owmsIdentifier": municipality.organisatie.owms_identifier,
-                "owmsPrefLabel": municipality.organisatie.owms_pref_label,
-                "owmsEndDate": municipality.organisatie.owms_end_date.isoformat(),
-                "locaties": [],
-                "catalogi": [],
-                "contactNaam": municipality.contact_naam,
-                "contactWebsite": municipality.contact_website,
-                "contactEmailadres": municipality.contact_emailadres,
-                "contactTelefoonnummer": municipality.contact_telefoonnummer,
-                "bevoegdeOrganisaties": [],
-                "ondersteuningsOrganisatie": {
-                    "owmsIdentifier": municipality.ondersteunings_organisatie.owms_identifier,
-                    "owmsPrefLabel": municipality.ondersteunings_organisatie.owms_pref_label,
-                    "owmsEndDate": municipality.ondersteunings_organisatie.owms_end_date.isoformat(),
-                },
-            },
-            data,
-        )
-
-
-class LocatiesTests(APITestCase):
-    def test_list_locations(self):
-        LocatieFactory.create_batch(2)
-        list_url = reverse("api:locatie-list")
-
-        response = self.client.get(list_url)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        data = response.json()["results"]
-        self.assertEqual(2, len(data))
-
-    def test_retrieve_location_by_uuid(self):
-        locatie = LocatieFactory.create()
-        detail_url = reverse("api:locatie-detail", args=[locatie.uuid])
-
-        response = self.client.get(detail_url)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        data = response.json()
-
-        self.assertEqual(str(locatie.uuid), data["uuid"])
